refactored censored message into component, removed autocensor message
